GameState Scene 1 Update
Work that needs to be done:
Figure out more of the outlineing + grabbale interactions. In particular, turn off knife XRScript + Outline Script during BeginBowlStep, then reactive these scripts during BeginBrownPork.

There is a bug with the bowl triggering BeginBrownPorkBook before all the objects are all in.  Should set up a debug tester and verify the number of objects meant to be in the bowl.

Should turn off the bowl OutlineScript when BeginBrownPork starts.

Should replace the block with the text book (whenever that may be done)

Need to ask which kitchen I should begin importing everything to.

For far future:
Need to figure out the spice packets.  Scissors will be found by team (presumably), and packet will be made.  Simply cut the packet with scissors and  learn how to do the pour animation.

Or you can just do the ladil scoop.

Small details that need to be fixed in mine:

(Probably do tomorrow) Actually do the pork in the pot

Ask about the sizzle sound as well.

Figure out more stuff to do after that.